+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Write instance URL to log at startup.
 - Change MSRV to 1.60.0.
+- Allow cross-origin requests from 127.0.0.1 by default.
 
 ## [2.4.0] - 2023-11-16